# Optimize the non-streaming XML parser