```
##### With input samples/hackernoon.xml #####
Name           ips        average  deviation         median         99th %
New         270.06        3.70 ms    ±37.15%        3.35 ms        8.60 ms
Old         204.78        4.88 ms    ±17.15%        4.67 ms        8.27 ms

Comparison:
New         270.06
Old         204.78 - 1.32x slower +1.18 ms

##### With input samples/reddit.xml #####
Name           ips        average  deviation         median         99th %
New         423.92        2.36 ms    ±18.53%        2.25 ms        3.94 ms
Old         387.29        2.58 ms    ±16.31%        2.47 ms        4.02 ms

Comparison:
New         423.92
Old         387.29 - 1.09x slower +0.22 ms

##### With input samples/soccer.xml #####
Name           ips        average  deviation         median         99th %
New          13.35       74.88 ms     ±9.83%       72.52 ms      110.51 ms
Old          10.77       92.82 ms     ±8.44%       91.53 ms      137.02 ms

Comparison:
New          13.35
Old          10.77 - 1.24x slower +17.94 ms
```